修改角色编辑按钮报错

+++ fhs_module/fhs_basics/fhs_basics_api/src/main/java/com/fhs/basics/service/UcenterMsRoleService.java
@@ -78,7 +78,7 @@ public interface UcenterMsRoleService extends BaseService<UcenterMsRoleVO, Ucent
      * @param roleId
      * @return
      */
-    String[] getRolePermissionButtons(Integer roleId);
+    String[] getRolePermissionButtons(Long roleId);
 
     /**
      * 根据角色查询按钮id
